Responsive layout for large screen. Rebase with upstream
