+++ src/styles/global.css
@@ -100,9 +100,14 @@
 }
 
 .layout-edit {
-  @apply bg-white w-full max-w-4xl mx-auto px-4 py-8 m-8  rounded-lg
+  @apply bg-white  max-w-4xl  2xl:max-w-screen-2xl mx-auto px-4 py-8 m-8  rounded-lg
 }
-
+@responsive {
+  .layout-edit-wide {
+    /* On wide screen put 2 forms side by side */
+    @apply max-w-full  flex flex-row justify-center gap-x-4 gap-y-0;
+  }
+}
 .md-h1 {
   @apply text-3xl font-medium tracking-tight mt-14 mb-4 !important;
 }
