+++ tailwind.config.js
@@ -15,6 +15,12 @@ module.exports = {
       fontFamily: {
         sans: ["ISO", ...defaultTheme.fontFamily.sans],
       },
+      colors: {
+        'custom-green': '#8BB191',//'#FF4500CC',
+        'custom-mango': '#FFB000',
+        'custom-teal': '#00A6A5',
+        'custom-mint': '#0DD3BB',
+      },
     },
   },
 };
